Complete Validations.

+++ src/main/java/edu/educate/dto/AttendanceDto.java
@@ -19,5 +19,5 @@ public class AttendanceDto extends BaseDto {
 
     private ElementDto element;
 
-    private Double grade;
+    private Float grade;
 }
